requirements.txt
gradio
opencv-python
numpy
matplotlib
Pillow
torch
# Special case for segment-anything: install directly from GitHub
git+https://github.com/facebookresearch/segment-anything.git